Removing the header files as suggested during review #11
