--- include/ut_test_runner.h
@@ -1,64 +1,0 @@
-/*
- * If not stated otherwise in this file or this component's LICENSE file the
- * following copyright and licenses apply:
- *
- * Copyright 2023 RDK Management
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-#ifndef UT_TEST_RUNNER_H
-#define UT_TEST_RUNNER_H
-
-#include <gtest/gtest.h>
-#include <functional>
-#include <string>
-
-class UTTestRunner
-{
-public:
-    explicit UTTestRunner(int &argc, char **argv)
-    {
-        ::testing::InitGoogleTest(&argc, argv);
-    }
-
-    void setTestFilter(const std::string &filter)
-    {
-        ::testing::GTEST_FLAG(filter) = filter;
-    }
-
-    void setGTestFlag(const std::string &flagName, const std::string &value)
-    {
-        if (flagName == "filter")
-        {
-            setTestFilter(value);
-        }
-        // Add more flags as needed
-    }
-
-    int runTests() const
-    {
-        return RUN_ALL_TESTS();
-    }
-
-    int runTestsWithCustomSetup(std::function<void()> setup = nullptr) const
-    {
-        if (setup)
-        {
-            setup();
-        }
-        return RUN_ALL_TESTS();
-    }
-};
-
-#endif // UT_TEST_RUNNER_H
